requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-pptx
openpyxl
Pillow
lxml
